refactor 2.0

+++ src/com/company/Graph/View/ConsoleViewWorker.java
@@ -1,7 +1,5 @@
 package com.company.Graph.View;
 
-import com.company.Graph.Model.AdjMatrix;
-
 import java.util.Scanner;
 
 public class ConsoleViewWorker {
@@ -37,7 +35,7 @@ public class ConsoleViewWorker {
         System.out.println(message);
     }
 
-    public static void printMatrixChar(double i) {
+    public static void printMatrixChar(int i) {
         System.out.print(i);
     }
